backend/requirements.txt
#fastapi[standard]==0.123.2
Django>=4.2
django-cors-headers
psycopg2-binary
django-extensions
pytest
django-import-export
django-summernote
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
premailer